# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Baseball Lineup Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List all games
      responses:
        "200":
          description: List of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      operationId: createGame
      tags: [games]
      summary: Create a new game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGameBody"
      responses:
        "201":
          description: Created game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
...
  /games/{id}/pitch-counts:
    get:
      operationId: getGamePitchCounts
      tags: [pitchCounts]
      summary: Get all pitch counts recorded for a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pitch counts for the game
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PitchCount"
    post:
      operationId: upsertGamePitchCount
      tags: [pitchCounts]
      summary: Upsert a player's pitch count for this game (one row per game+player)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertPitchCountBody"
      responses:
        "200":
          description: Upserted pitch count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PitchCount"
  /games/{gameId}/pitch-counts/{playerId}:
    delete:
      operationId: deleteGamePitchCount
      tags: [pitchCounts]
      summary: Delete a player's pitch count for this game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "409":
          description: Row version conflict — the record changed on another device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
    ConflictError:
      type: object
      description: Returned with HTTP 409 when an If-Match row version no longer matches — another device changed the record. `current` is the row as it exists now.
      properties:
        error:
          type: string
        code:
          type: string
          enum: [row_version_conflict, idempotency_in_flight]
        current: {}
      required: [error, code]
...
    Game:
      type: object
      properties:
        rowVersion:
          type: integer
          description: Optimistic-concurrency version, incremented on every update. Send back in If-Match on PATCH/DELETE to detect conflicts.
        id:
          type: integer
        opponent:
          type: string
        gameDate:
          type: string
          format: date-time
        location:
          type: ["string", "null"]
        innings:
          type: integer
          description: Number of innings in the game
        status:
          type: string
          enum: [upcoming, completed, cancelled]
        type:
          type: string
          enum: [game, practice, other]
          description: Event kind — distinguishes games from practices and other team events
        gameType:
          type: ["string", "null"]
          enum: [league, tournament, null]
          description: Competitive context for lineup generation — "tournament" favors competitiveness, "league" rebalances season plate appearances, null falls back to the global equity slider
        competitiveness:
          type: ["integer", "null"]
          minimum: 0
          maximum: 100
          description: Per-game competitiveness override (0-100) for the batting order. 0 = even out plate appearances, 100 = best OPS order blended with the coach's learned preferred slots. Null falls back to the gameType/global fairness behavior.
        ourScore:
          type: ["integer", "null"]
        opponentScore:
          type: ["integer", "null"]
        startedAt:
          type: ["string", "null"]
          format: date-time
          description: Wall-clock time of the actual first pitch (coach taps "Start Game" on the field display). Null until kicked off. Distinct from gameDate (scheduled start) — used by the running game timer.
        tournamentId:
          type: ["integer", "null"]
          description: Optional FK to a tournaments row. Set when the game is part of a multi-game tournament weekend so the app can roll per-pitcher pitch counts across the tournament.
        bracketStage:
          type: ["string", "null"]
          enum: [pool, bracket, null]
          description: For tournament games — which stage this game belongs to. Drives which time-limit pair applies. Null when the game has no tournamentId.
        isChampionship:
          type: boolean
          description: True when this game is flagged as the tournament championship. The Field Display auto-enables Championship Mode for it. Only ever true on tournament games.
        effectiveTimeLimits:
          oneOf:
            - type: "null"
            - type: object
              properties:
                noNewInningMinutes:
                  type: ["integer", "null"]
                hardStopMinutes:
                  type: ["integer", "null"]
              required: [noNewInningMinutes, hardStopMinutes]
          description: Server-resolved time-limit rules for this game (looked up from the parent tournament's pool vs bracket fields based on `bracketStage`). Null when the game has no `tournamentId` or the resolved pair is fully empty.
        notes:
          type: ["string", "null"]
        planSnapshot:
          oneOf:
            - type: "null"
            - type: array
              items:
                $ref: "#/components/schemas/PlanSnapshotEntry"
          description: Snapshot of the planned lineup taken before a post-game photo override (null when no snapshot has been taken)
        boxScoreImportedAt:
          type: ["string", "null"]
          format: date-time
          description: Timestamp of the most recent box-score import. Null if no box score has been imported. The UI shows an "Already imported on …" banner when set.
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - opponent
        - gameDate
        - innings
        - status
        - type
        - isChampionship
        - createdAt
    PlanSnapshotEntry:
      type: object
      properties:
        playerId:
          type: integer
        playerName:
          type: string
        inning:
          type: integer
        position:
          type: string
        battingOrder:
          type: ["integer", "null"]
      required:
        - playerId
        - playerName
        - inning
        - position
        - battingOrder
    CreateGameBody:
      type: object
      properties:
        opponent:
          type: string
        gameDate:
          type: string
          format: date-time
        location:
          type: ["string", "null"]
        innings:
          type: integer
        notes:
          type: ["string", "null"]
        gameType:
          type: ["string", "null"]
          enum: [league, tournament, null]
        competitiveness:
          type: ["integer", "null"]
          minimum: 0
          maximum: 100
          description: Per-game competitiveness override (0-100) for the batting order. Null = use the gameType/global fairness behavior.
        tournamentId:
          type: ["integer", "null"]
          description: Optionally create the game already linked to a tournament.
      required:
        - opponent
        - gameDate
        - innings
...
    PitchCount:
      type: object
      properties:
        rowVersion:
          type: integer
          description: Optimistic-concurrency version, incremented on every update. Send back in If-Match on PATCH/DELETE to detect conflicts.
        id:
          type: integer
        gameId:
          type: integer
        playerId:
          type: integer
        pitches:
          type: integer
          minimum: 0
        notes:
          type: ["string", "null"]
        recordedAt:
          type: string
          format: date-time
      required:
        - id
        - gameId
        - playerId
        - pitches
        - recordedAt
    UpsertPitchCountBody:
      type: object
      properties:
        playerId:
          type: integer
        pitches:
          type: integer
          minimum: 0
          maximum: 500
        notes:
          type: ["string", "null"]
      required:
        - playerId
        - pitches
...